chisel-crv: finish implementation with BigInt
I'm not very happy with the current implementation, I would like to
create a class that is more natura to use. Have a simple DSL that can
provide more functionalities, like I would like to have a "fake"
assign operator or at least have a decent implicit method that
transform my integers in actual Rand variables. In order to do so, I
need to change the current implemntation of the Module and how we are
currently assign constraint etc. Probably create a model every time we
call the randomize function will do the trick or probably not. I need
a way to register the value created by the implict inside the model.